+++ localizations/en_us/localizations.lua
@@ -1,0 +1,7 @@
+local localizations = {}
+
+localizations.values = {
+    ui_action_back = "Back", 
+}
+
+return localizations