+++ scripts/hammerstone/utils/shadow.lua
@@ -1,6 +1,9 @@
 -- Hammerstone: shadow.lua
 -- @author SirLich
 
+-- Hammerstone
+local logging = mjrequire "hammerstone/logging"
+
 local shadow = {}
 
 -- Prepares a file to be shadowed by the base game.
@@ -16,6 +19,15 @@ function shadow:shadow(outerModule, loadOrder)
 	-- @param parentModule is the actual base file from Sapiens
 	-- @param outerModule is the current outer module (shadow), passed into shadow:shadow
 	function outerModule:onload(parentModule)
+
+		-- if the parentModule is a string, it means an error occured during "require"
+		-- (probably due to patching)
+		-- the string is the error message returned by pcall(require, ...) in the vanilla code
+		-- of common/modManager
+		if type(parentModule) == "string" then
+			logging:error("Error reading base game file:", parentModule)
+			return
+		end
 
 		if outerModule.preload ~= nil then
 			outerModule:preload(parentModule)